hashes: allocate 16 more bytes for the buffer
We'll print a long long unsigned int later in MAGIC_PE64.
Which will, theoretically, yields up to 20 bytes in some platforms.
Hence, gcc will report a warning for that snprintf.

In practice, we will never run into that situation,
but 16 bytes is not that expensive.

Let's play an easy game.